move constants under /app
