+++ eas-indexer/src/api/index.ts
@@ -4,7 +4,6 @@ import * as dbSchema from "../../ponder.schema";
 import schemas from "../../schemas.config";
 import { Attestation, Entity, AggregatedType } from "../types";
 import { parseEntity } from "./utils";
-import { COMMUNITY_CONTRIBUTORS_ATTEST_ADDRESS } from "../constants";
 
 ponder.use("/", graphql());
 
@@ -56,37 +55,6 @@ ponder.get("/entities/aggregated", async (c) => {
 
     aggregated[entity] = attestations;
   }
-
-  // for (const entity of entities) {
-  //   const table = dbSchema[entity] as any;
-  //   const data = await (c.db.query[entity] as any).findMany({
-  //     where: isNull(table.revoked_at),
-  //   });
-
-  // if (entity === "rf_voter") {
-  //   data = await (c.db.query[entity] as any).findMany({
-  //     where: and(isNull(table.revoked_at), eq(table.voter_type, "rf_voter")),
-  //   });
-  // } else if (entity === "gov_contribution") {
-  //   data = await (c.db.query[entity] as any).findMany({
-  //     where: and(isNull(table.revoked_at), eq(table.gov_season, "7")),
-  //   });
-  // } else {
-  //   data = await (c.db.query[entity] as any).findMany({
-  //     where: isNull(table.revoked_at),
-  //   });
-  // }
-
-  // aggregated[entity] = data
-
-  // aggregated.community_contributors = data.reduce((acc: any, item: any) => {
-  //   if (item.attester === COMMUNITY_CONTRIBUTORS_ATTEST_ADDRESS) {
-  //     acc.push({ address: item.address });
-  //   }
-  //   return acc;
-  // }, []);
-  // }
-  // }
 
   return c.json(aggregated);
 });
