--- eas-indexer/src/constants.ts
@@ -1,87 +1,0 @@
-import { AggregatedType } from "./types";
-
-export const TAG_BY_ENTITY: Record<keyof AggregatedType, string> = {
-  citizen: "Citizen",
-  community_contributors: "Community Contributor",
-  onchain_builders: "Onchain Builder",
-  badgeholder: "Badgeholder",
-  gov_contribution: "S7 Elected Official",
-  rf_voter: "Guest Voter",
-};
-
-export const COMMUNITY_CONTRIBUTORS_ATTEST_ADDRESS =
-  "0xc975707485575a611b1aaf9cf9aada04a066b690e40aec7eb5cb17e22336d703";
-
-export const CONTRIBUTOR_ELIGIBLE_PROJECTS = [
-  "0x000c2ce4773defb3010a58d3800d0ec9d432189c574ba26d3103f1db44a5af6d",
-  "0x04b1cd5a7c59117474ce414b309fa48e985bdaab4b0dab72045f74d04ebd8cff",
-  "0x08df6e20a3cfabbaf8f34d4f4d048fe7da40447c24be0f3ad513db6f13c755dd",
-  "0x09b95c7697625da4915338750c5f78446817a3634cb38bc9155e26bbbc0c87f1",
-  "0x0fad18f37ecec4f4f4ca18e60c353eb11462363e8aa14562b5d96af54108d106",
-  "0x182f5a5c04ccdc4739ee9f362af72d7a0f55d6b410cc9dc2b52dc7f99f06f3f9",
-  "0x1a6ff49fed19e164fd6b39f668f391cfc7fc493906e942c74997aacee91be632",
-  "0x1ea2fc7183a8ad244a97ecc287ca922209274c11dd9e991a04315e8399e118fc",
-  "0x2e075fb37977668c15af474229f2f2abb009850a6a867b8bbf4adb5e293d1fe3",
-  "0x3061b642db56c507e265f03029735b0413a613bda43456a31e78128baeb18754",
-  "0x3230536d9c471231698a0fe66daeeda963282f2b594ce5a47d3dbfa53ac9d4a5",
-  "0x32942ffa0d751769bb55b53d9c32085463e9bfe668ae8aa5d4a736c46b4130c5",
-  "0x3a103436d12a3d54ade1a44f87ce255417fd530b6ec55abd3bea1898b32c61c1",
-  "0x3ecbf66c1b21342b26acb0f7d4c6b8bbbef168ffafff3f425de3503fd8dfa535",
-  "0x41b2ae8a2ead458ff7c85549a9a11380353e1ae4b234666312d495326636f77a",
-  "0x4288aacdda65af69c03b2c9b9d057e2b93ed1b7671e481b03c6a7b33349d2365",
-  "0x4cf98b5dec7f0f51eb6b88ecb15639bafc7320bdc6e21ac443f4c58ebeb1245b",
-  "0x517eaa9c56951de89261f2d7830ea49aae92f2a903104a17d9c5c2edd4959806",
-  "0x5759249c433d67eeb2ca1b6ff827feec164b60b92e849d6ce0db0974cedc4a89",
-  "0x5c4cbf1cf48fea0f82cfdf33b94f43910f3062b6cca1fbfe46d4ee7e1f5506d1",
-  "0x5e273cdfe70177d7e48afcc29f2cc76a361d9bc49edf294d72089d2fed156871",
-  "0x5e6c436e48e56d6d9622ba5d0be0035c314e2b29d2afc8f5f1ee8ac75cd42532",
-  "0x5f1def93f7d757bd498e7766d3b3735f5fd899c48ddda039b5e23b73a95fa4ec",
-  "0x60f8ab8d1ae00182f9afaf6ec35213a200ed8ac21d39b24a9794f2ba21f6aa59",
-  "0x62e37e96aa6e1cbfb6bd24b97c4b8f1e12cc3fe35d5388d2f041c42a12b40745",
-  "0x6830f24e2e893931d62d8118146a7bbffb5cd1f78945318865a52bd6a6cbe245",
-  "0x684a9b5fbc38b6577950f0c641631e190c6edafc67d00ba6d8334a87ac95b407",
-  "0x69f13f6351489812a9052412e273ba726383744c8c482ac877c6a4e9f3b1d3ad",
-  "0x6aa80764b6082947c3b2de86fe12804eb475b0afb719de50b9eed60b86f20535",
-  "0x6dab3b38a62faa0195c0a989aca19c1b11e8ebbc97cf9d7269018850e0606550",
-  "0x70a31809b59034a137b8ba57b092412879e779f0ea527f091776dca521edc7af",
-  "0x7539f43dc91f9c406c83a06339708e4033cdbf630916c6dfb9de2b5ef811affd",
-  "0x77cc2b19e794565ceacafa96ad5fdc27964b605296e0a2303acc93d12659891d",
-  "0x78321a16ccf90916db134d71eb0aca96eed5a4845695700a08427ba77bb8fa43",
-  "0x79c2ae8858a5a3bd7f2e0da9509d727833ffa01f358b5969355f38ff32b20122",
-  "0x81af86360a8e964cc907a5689a62e94231c7e42e05f49c0586716975d72e2e2e",
-  "0x81f3e812a5677ddc5fb1a1e0c38ce6eb03f3b0a11a65283f42cdf93ceb5f66f8",
-  "0x8431c44e2f0903879f34134e68fba4c05833b02451ccbf4cdb69620a6715bcdd",
-  "0x850df51e29f2846a5f085d88e6b6fc13fad51ad7161f8a825ed09d470668161a",
-  "0x8e60d35b5b545cfb3a2264c6f95024c0359b00a3febcb7a56967934ce917783c",
-  "0x9056aef66264e926d4aefa66cf45438b2586459934515dff43432e7a95d0d972",
-  "0x9102357674825ed224734387fbefbf41c16fe5d9764c2f7f3e897ad3325d3990",
-  "0x91a4420e2fcc8311e97dad480f201a8ce221f2cd64c2de77280cbcc6ce193752",
-  "0x945e356d400d37b4cd548ca3b53d4ee41a075c8b4befd0b8589947b6d8721ad4",
-  "0x9ca1f7b0e0d10d3bd2619e51a54f2e4175e029c87a2944cf1ebc89164ba77ea0",
-  "0x9fda488eb998f9c97fde115f2145be4eadad954fa539adda0afb21be5c04e829",
-  "0xa38f3efb4fb8f6fcefb80f0262645ac05d5548cad0308ee49520c48c4e8cbd1f",
-  "0xa88844cea135382e3484e39c3172033437121b35ca0bc8b10b9b8253984876b5",
-  "0xb2d109759fe14e11ac5cc100ab6006321ebdd7ffdefbd2efac93a002105f8e92",
-  "0xb98778ca9ff41446e2bc304f7b5d27f0fa7c2bcd11df19e22d1352c06698a1f6",
-  "0xbb3291f2e0a56c4f8fe8250613c5487e4447fce05bb75854472fe3e01eebaffa",
-  "0xbcb5c437d5347be21add3972512679054416805274f80e85ab8053ae79340a69",
-  "0xc49d46c5606148617c4202a6a0f745ce194418ca81f1c27d1ff75ea84e32a0cc",
-  "0xc71faa1bcb4ceb785816c3f22823377e9e5e7c48649badd9f0a0ce491f20d4b3",
-  "0xc879d4a2a344a4027ee8abd520d65afdfe1300b6f080b223fafcebdf5640c218",
-  "0xcbfdcf8f724d73504d1518d5ecd7d7256b12d37e68cd84569b383bfb6e2eeea4",
-  "0xcc8d03e014e121d10602eeff729b755d5dc6a317df0d6302c8a9d3b5424aaba8",
-  "0xcd9dfcace81ae28b29b10c4ddf36bcb8db6d705cbb92e7d885e5b946a9f77910",
-  "0xcf10a15439ef8c56e1b6c22641544012bfd032c7b584d1160a7d743cb770875b",
-  "0xd4ed99cc6aaf73ca63b32f7a03b5427ac1d2955bf9efc31eb14f5773016988d0",
-  "0xd54e1032bf16d649f8f8f0a7da42a249966ad4a270b4c17b24ded1e32c97b2bf",
-  "0xd5b4c54b12bf86f6eb67fec81032809a16ff1c4b4c7f0d5898fc86367db86ca6",
-  "0xda38e275285330082433713e32b03c7ccd482bfca5e52719766f16555e05f699",
-  "0xe15a32ba713cc025574ac6f69be3ad16133fb7561c3d7dc30fbf29b65b5a8be2",
-  "0xe2247097961708233325f253d676ba2f9d8011147e1bb08e0e50d55894138dc7",
-  "0xe2f742ea351ea741e95f8cf4b2434317d8365c532b97601e8a27854003c2e396",
-  "0xe346264e87202b47f1057eb0b0fcaa0ea7f83e14507ca4585a91a5d94e0e92c0",
-  "0xefaeef1446de875ab6e50d31c8dbe776f99d884cb8cb68d50b1384067e4aeb6b",
-  "0xf24315614063278ba3543e1717791132a9afa79b0e65baa01a85bcccbdfa215f",
-  "0xfad78fad680d407f81e7fd46632b4b4936676d1776a5cbf02694b94698495746",
-  "0xffffeb1d3576c1e847701c225975ee1a10259e05366b477e98fea7a47c5e5896",
-];
